refactor: rename rout to navigation and move it under core directly
