+++ lib/core/navigate/app_navigator_impl.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:go_router/go_router.dart';
-import 'package:sellio_mobile/core/app_management/route/route_args.dart';
+import 'package:sellio_mobile/core/navigate/route_args.dart';
 
 import 'app_navigator.dart';
 import 'app_routes.dart';
