+++ lib/core/navigate/route_manager.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:sellio_mobile/core/app_management/route/route_args.dart';
+import 'package:sellio_mobile/core/navigate/route_args.dart';
 import '../../../presentation/screens/auth/create_account/create_account_screen.dart';
 import '../../../presentation/screens/auth/forgetPassword/confirm_password_screen.dart';
 import '../../../presentation/screens/auth/forgetPassword/forget_password_otp_screen.dart';
