+++ lib/presentation/screens/auth/create_account/create_account_listeners.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:sellio_mobile/core/app_management/route/routing.dart';
+import 'package:sellio_mobile/core/navigate/routing.dart';
 
 import '../../../../core/design_system/widgets/snack_bar.dart';
 import 'cubits/form/create_account_form_cubit.dart';
