+++ lib/presentation/screens/auth/create_account/widgets/create_account_footer.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../../core/app_management/route/routing.dart';
+import '../../../../../core/navigate/routing.dart';
 import '../../../../../core/design_system/constants/app_strings.dart';
 import '../../../../../core/design_system/themes/sellio_theme_provider.dart';
