+++ lib/presentation/screens/auth/login/widgets/login_form.dart
@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gap/flutter_gap.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:sellio_mobile/core/app_management/route/routing.dart';
+import 'package:sellio_mobile/core/navigate/routing.dart';
 
 import '../../../../../core/design_system/constants/app_strings.dart';
 import '../../../../../core/design_system/constants/assets.dart';
