+++ lib/presentation/screens/auth/signupOTP.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:sellio_mobile/core/app_management/route/navigation_extensions.dart';
+import 'package:sellio_mobile/core/navigate/navigation_extensions.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
 import 'package:sellio_mobile/core/design_system/widgets/cards/otp_card.dart';
 import 'package:sellio_mobile/core/localization/localization_service.dart';
