+++ lib/presentation/screens/home/utils/home_navigation.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../../../core/app_management/route/navigation_extensions.dart';
+import '../../../../core/navigate/navigation_extensions.dart';
 import '../../../../domain/entities/store.dart';
 import '../../store_details/store_details_screen.dart';
